[RENAME] CachSync디렉토리 이동
